% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mvshape.R
\name{floatvar}
\alias{floatvar}
\title{floatvar}
\usage{
floatvar(V, tol = 0.001, iter.max = 50)
}
\arguments{
\item{tol}{tolerance parameter.}

\item{iter.max}{maximum number of iterations.}

\item{v}{variance matrix.}
}
\value{
List of floating point variances, error limits and divergence parameters.
}
\description{
floatvar computes floating point variances.
}
\author{
James R Staley <jrstaley95@gmail.com>
}
