% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mvshape.R
\name{fracpoly_plot}
\alias{fracpoly_plot}
\title{fracpoly_plot}
\usage{
fracpoly_plot(fracpoly, degree = "both", xref = NULL, logx = FALSE,
  logy = FALSE, pref_x = "x", pref_y = "y", xbreaks = NULL,
  ybreaks = NULL, cicolour = "grey", betacolour = "black",
  intcolour = "grey", xlim = NULL)
}
\arguments{
\item{fracpoly}{a fracpoly object.}

\item{degree}{the fractonal polyniomal degree to be plotted (options: 1, 2 and both).}

\item{xref}{the reference point for the figures of binary outcomes.}

\item{logx}{plot the x-axis on the log-scale.}

\item{logy}{plot the y-axis on the log-scale.}

\item{pref_x}{the prefix for the x-axis.}

\item{pref_y}{the prefix for the y-axis.}

\item{xbreaks}{breaks in the x-axis.}

\item{ybreaks}{breaks in the y-axis.}

\item{cicolour}{colour of the 95\% confidence intervals.}

\item{betacolour}{colour of the regression line.}

\item{intcolour}{colour of the intercept line.}

\item{xlim}{x-axis limits.}
}
\value{
fractional polynomial plot.
}
\description{
fracpoly_plot plots the best fitting fractional polynomial.
}
\examples{
### Data
y <- rnorm(5000)
x <- rnorm(5000,10,1)
c1 <- rbinom(5000,1,0.5)
c2 <- rnorm(5000)
covar <- data.frame(c1=c1, c2=c2)

### Analyses
res <- fracpoly(y=y, x=x, covar=covar, family="gaussian")
fracploy_plot(res)
}
\author{
James R Staley <jrstaley95@gmail.com>
}
