% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mvshape.R
\name{mvshape}
\alias{mvshape}
\title{mvshape}
\usage{
mvshape(y = y, x = x, covar = NULL, study = NULL, ngrp = 10,
  refgrp = 1, family = "gaussian", float = FALSE, method = "reml")
}
\arguments{
\item{y}{outcome.}

\item{x}{exposure.}

\item{covar}{data.frame with covariates.}

\item{study}{study variable.}

\item{ngrp}{number of quantiles of the exposure.}

\item{refgrp}{reference group.}

\item{family}{the glm family (options: gaussian and binomial).}

\item{float}{floating point variances.}

\item{method}{meta-analysis method.}
}
\value{
List of multivariate meta-analysis results for each group.
}
\description{
mvshape fits a multivariate meta-analysis for groups of the exposure (e.g. deciles).
}
\examples{
### Data
y <- rnorm(5000)
x <- rnorm(5000,10,1)
c1 <- rbinom(5000,1,0.5)
c2 <- rnorm(5000)
covar <- data.frame(c1=c1, c2=c2)
study <- c(rep("study1",1000),rep("study2",1000),rep("study3",1000),rep("study4",1000),rep("study5",1000))

### Analyses
res <- mvshape(y=y, x=x, covar=covar, study=study, family="gaussian")
}
\author{
James R Staley <js16174@bristol.ac.uk>
}
