% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mvshape.R
\name{mvshape_plot}
\alias{mvshape_plot}
\title{mvshape_plot}
\usage{
mvshape_plot(
  mvshape,
  logx = FALSE,
  logy = FALSE,
  pref_x = "x",
  pref_y = "y",
  xbreaks = NULL,
  ybreaks = NULL,
  betacolour = "red",
  cicolour = "red",
  intcolour = "grey",
  xlim = NULL
)
}
\arguments{
\item{mvshape}{a mvshape object.}

\item{logx}{plot the x-axis on the log-scale.}

\item{logy}{plot the y-axis on the log-scale.}

\item{pref_x}{the prefix for the x-axis.}

\item{pref_y}{the prefix for the y-axis.}

\item{xbreaks}{breaks in the x-axis.}

\item{ybreaks}{breaks in the y-axis.}

\item{betacolour}{colour of the regression line.}

\item{cicolour}{colour of the 95\% confidence intervals.}

\item{intcolour}{colour of the intercept line.}

\item{xlim}{x-axis limits.}
}
\value{
mvshape plot.
}
\description{
mvshape_plot plots the meta-analysis results against the mean values of the exposure in each group.
}
\examples{
# Data
y <- rnorm(5000)
x <- rnorm(5000, 10, 1)
c1 <- rbinom(5000, 1, 0.5)
c2 <- rnorm(5000)
study <- c(rep("study1", 1000), rep("study2", 1000), rep("study3", 1000), rep("study4", 1000), rep("study5", 1000))
covar <- data.frame(c1 = c1, c2 = c2, study = study)

# Analyses
res <- mvshape(y = y, x = x, covar = covar[, c("c1", "c2")], study = study, family = "gaussian")
mvshape_plot(res)
}
\author{
James Staley \href{mailto:jrstaley95@gmail.com}{jrstaley95@gmail.com}
}
