% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mvshape.R
\name{print.fracpoly}
\alias{print.fracpoly}
\title{Print fracpoly}
\usage{
\method{print}{fracpoly}(x, ...)
}
\arguments{
\item{x}{an object of class \code{"fracpoly"}.}
}
\description{
print method for class \code{"fracpoly"}.
}
\author{
James Staley \href{mailto:jrstaley95@gmail.com}{jrstaley95@gmail.com}
}
