% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mvshape.R
\name{print.mvshape}
\alias{print.mvshape}
\title{Print mvshape}
\usage{
\method{print}{mvshape}(x, ...)
}
\arguments{
\item{x}{an object of class \code{"mvshape"}.}
}
\description{
print method for class \code{"mvshape"}.
}
\author{
James Staley \href{mailto:jrstaley95@gmail.com}{jrstaley95@gmail.com}
}
