% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mvshape.R
\name{summary.fracpoly}
\alias{summary.fracpoly}
\title{Summary fracpoly}
\usage{
\method{summary}{fracpoly}(x, ...)
}
\arguments{
\item{x}{an object of class \code{"fracpoly"}.}
}
\description{
summary method for class \code{"fracpoly"}.
}
\author{
James Staley \href{mailto:jrstaley95@gmail.com}{jrstaley95@gmail.com}
}
