% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mvshape.R
\name{fracpoly}
\alias{fracpoly}
\title{fracpoly}
\usage{
fracpoly(y = y, x = x, covar = NULL, family = "gaussian")
}
\arguments{
\item{y}{outcome.}

\item{x}{exposure.}

\item{covar}{data.frame with covariates.}

\item{family}{the glm family (options: gaussian and binomial).}
}
\value{
List of best fitting polynomials of degrees 1 and 2 as well as associated statistics.

\item{power_d1}{power of the best-fitting fractional polynomial of degree 1}

\item{fp1}{model of the best-fitting fractional polynomial of degree 1}

\item{power_d2}{powers of the best-fitting fractional polynomial of degree 2}

\item{fp2}{model of the best-fitting fractional polynomial of degree 2}

\item{p_d1}{p-value testing the best-fitting fractional polynomial of degree 1 against the linear model}

\item{p_d2}{p-value testing the best-fitting fractional polynomial of degree 2 against the best-fitting fractional polynomial of degree 2}

\item{xmin}{miniumum value of the exposure}

\item{xmax}{maximum value of the exposure}

\item{family}{family used in the analysis}
}
\description{
fracpoly fits the best fitting fractional polynomial of degree 1 and 2.
}
\examples{
### Data
y <- rnorm(5000)
x <- rnorm(5000,10,1)
c1 <- rbinom(5000,1,0.5)
c2 <- rnorm(5000)
covar <- data.frame(c1=c1, c2=c2)

### Analyses
res <- fracpoly(y=y, x=x, covar=covar, family="gaussian")
}
\author{
James R Staley <jrstaley95@gmail.com>
}
