% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mvshape.R
\name{print.fracpoly}
\alias{print.fracpoly}
\title{Print fracpoly}
\usage{
\method{print}{fracpoly}(x, ...)
}
\arguments{
\item{x}{an object of class "fracpoly".}
}
\description{
print method for class "fracpoly".
}
\author{
James R Staley <jrstaley95@gmail.com>
}
