% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mvshape.R
\name{print.mvshape}
\alias{print.mvshape}
\title{Print mvshape}
\usage{
\method{print}{mvshape}(x, ...)
}
\arguments{
\item{x}{an object of class "mvshape".}
}
\description{
print method for class "mvshape".
}
\author{
James R Staley <jrstaley95@gmail.com>
}
