% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mvshape.R
\name{summary.fracpoly}
\alias{summary.fracpoly}
\title{Summary fracpoly}
\usage{
\method{summary}{fracpoly}(x, ...)
}
\arguments{
\item{x}{an object of class "fracpoly".}
}
\description{
summary method for class "fracpoly".
}
\author{
James R Staley <js16174@bristol.ac.uk>
}
