% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mvshape.R
\name{summary.mvshape}
\alias{summary.mvshape}
\title{Summary mvshape}
\usage{
\method{summary}{mvshape}(x, ...)
}
\arguments{
\item{x}{an object of class "mvshape".}
}
\description{
summary method for class "mvshape".
}
\author{
James R Staley <jrstaley95@gmail.com>
}
